+++ integration_tests/codegen_fail/fail/scalar/derive_invalid_url.rs
@@ -1,0 +1,7 @@
+use juniper::GraphQLScalarValue;
+
+#[derive(GraphQLScalarValue)]
+#[graphql(specified_by_url = "not an url")]
+struct ScalarSpecifiedByUrl(i64);
+
+fn main() {}