+++ juniper/src/schema/model.rs
@@ -1,4 +1,4 @@
-use std::fmt;
+use std::{borrow::Cow, fmt};
 
 use fnv::FnvHashMap;
 #[cfg(feature = "graphql-parser-integration")]
@@ -49,6 +49,7 @@ pub struct RootNode<
 /// Metadata for a schema
 #[derive(Debug)]
 pub struct SchemaType<'a, S> {
+    pub(crate) description: Option<Cow<'a, str>>,
     pub(crate) types: FnvHashMap<Name, MetaType<'a, S>>,
     pub(crate) query_type_name: String,
     pub(crate) mutation_type_name: Option<String>,
@@ -71,6 +72,7 @@ pub struct DirectiveType<'a, S> {
     pub description: Option<String>,
     pub locations: Vec<DirectiveLocation>,
     pub arguments: Vec<Argument<'a, S>>,
+    pub is_repeatable: bool,
 }
 
 #[derive(Clone, PartialEq, Eq, Debug, GraphQLEnum)]
@@ -80,12 +82,19 @@ pub enum DirectiveLocation {
     Mutation,
     Subscription,
     Field,
+    Scalar,
     #[graphql(name = "FRAGMENT_DEFINITION")]
     FragmentDefinition,
+    #[graphql(name = "FIELD_DEFINITION")]
+    FieldDefinition,
+    #[graphql(name = "VARIABLE_DEFINITION")]
+    VariableDefinition,
     #[graphql(name = "FRAGMENT_SPREAD")]
     FragmentSpread,
     #[graphql(name = "INLINE_FRAGMENT")]
     InlineFragment,
+    #[graphql(name = "ENUM_VALUE")]
+    EnumValue,
 }
 
 impl<'a, QueryT, MutationT, SubscriptionT>
@@ -211,6 +220,14 @@ impl<'a, S> SchemaType<'a, S> {
             "include".to_owned(),
             DirectiveType::new_include(&mut registry),
         );
+        directives.insert(
+            "deprecated".to_owned(),
+            DirectiveType::new_deprecated(&mut registry),
+        );
+        directives.insert(
+            "specifiedBy".to_owned(),
+            DirectiveType::new_specified_by(&mut registry),
+        );
 
         let mut meta_fields = vec![
             registry.field::<SchemaType<S>>("__schema", &()),
@@ -235,6 +252,7 @@ impl<'a, S> SchemaType<'a, S> {
             }
         }
         SchemaType {
+            description: None,
             types: registry.types,
             query_type_name,
             mutation_type_name: if &mutation_type_name != "_EmptyMutation" {
@@ -249,6 +267,11 @@ impl<'a, S> SchemaType<'a, S> {
             },
             directives,
         }
+    }
+
+    /// Add a description.
+    pub fn set_description(&mut self, description: impl Into<Cow<'a, str>>) {
+        self.description = Some(description.into());
     }
 
     /// Add a directive like `skip` or `include`.
@@ -489,12 +512,14 @@ where
         name: &str,
         locations: &[DirectiveLocation],
         arguments: &[Argument<'a, S>],
+        is_repeatable: bool,
     ) -> DirectiveType<'a, S> {
         DirectiveType {
             name: name.to_owned(),
             description: None,
             locations: locations.to_vec(),
             arguments: arguments.to_vec(),
+            is_repeatable,
         }
     }
 
@@ -510,6 +535,7 @@ where
                 DirectiveLocation::InlineFragment,
             ],
             &[registry.arg::<bool>("if", &())],
+            false,
         )
     }
 
@@ -525,6 +551,34 @@ where
                 DirectiveLocation::InlineFragment,
             ],
             &[registry.arg::<bool>("if", &())],
+            false,
+        )
+    }
+
+    fn new_deprecated(registry: &mut Registry<'a, S>) -> DirectiveType<'a, S>
+    where
+        S: ScalarValue,
+    {
+        Self::new(
+            "deprecated",
+            &[
+                DirectiveLocation::FieldDefinition,
+                DirectiveLocation::EnumValue,
+            ],
+            &[registry.arg::<String>("reason", &())],
+            false,
+        )
+    }
+
+    fn new_specified_by(registry: &mut Registry<'a, S>) -> DirectiveType<'a, S>
+    where
+        S: ScalarValue,
+    {
+        Self::new(
+            "specifiedBy",
+            &[DirectiveLocation::Scalar],
+            &[registry.arg::<String>("url", &())],
+            false,
         )
     }
 
@@ -536,24 +590,28 @@ where
 
 impl fmt::Display for DirectiveLocation {
     fn fmt(&self, f: &mut fmt::Formatter) -> fmt::Result {
-        f.write_str(match *self {
-            DirectiveLocation::Query => "query",
-            DirectiveLocation::Mutation => "mutation",
-            DirectiveLocation::Subscription => "subscription",
-            DirectiveLocation::Field => "field",
-            DirectiveLocation::FragmentDefinition => "fragment definition",
-            DirectiveLocation::FragmentSpread => "fragment spread",
-            DirectiveLocation::InlineFragment => "inline fragment",
+        f.write_str(match self {
+            Self::Query => "query",
+            Self::Mutation => "mutation",
+            Self::Subscription => "subscription",
+            Self::Field => "field",
+            Self::FieldDefinition => "field definition",
+            Self::FragmentDefinition => "fragment definition",
+            Self::FragmentSpread => "fragment spread",
+            Self::InlineFragment => "inline fragment",
+            Self::VariableDefinition => "variable definition",
+            Self::Scalar => "scalar",
+            Self::EnumValue => "enum value",
         })
     }
 }
 
 impl<'a, S> fmt::Display for TypeType<'a, S> {
     fn fmt(&self, f: &mut fmt::Formatter) -> fmt::Result {
-        match *self {
-            TypeType::Concrete(t) => f.write_str(t.name().unwrap()),
-            TypeType::List(ref i, _) => write!(f, "[{}]", i),
-            TypeType::NonNull(ref i) => write!(f, "{}!", i),
+        match self {
+            Self::Concrete(t) => f.write_str(t.name().unwrap()),
+            Self::List(i, _) => write!(f, "[{}]", i),
+            Self::NonNull(i) => write!(f, "{}!", i),
         }
     }
 }